+++ feed.html
@@ -66,7 +66,7 @@
             placeholder="Write something about your day!"
           ></textarea>
           <div class="make-post-features w-100 d-flex justify-content-end">
-            <button class="dark-button">Post</button>
+            <button class="submit-post-button dark-button">Post</button>
           </div>
         </form>
       </div>
